refactor: move base rollup logic outside of the IndexedTree impl
