+++ yarn-project/merkle-tree/src/indexed_tree/indexed_tree.ts
@@ -27,21 +27,6 @@ export interface LeafData {
   nextValue: bigint;
 }
 
-export interface LowNullifierWitnessData {
-  /**
-   * Preimage of the low nullifier that proves non membership
-   */
-  preimage: LeafData;
-  /**
-   * Sibling path to prove membership of low nullifier
-   */
-  siblingPath: SiblingPath;
-  /**
-   * The index of low nullifier
-   */
-  index: bigint;
-}
-
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const encodeTreeValue = (leafData: LeafData) => {
   const valueAsBuffer = toBufferBE(leafData.value, 32);
@@ -117,11 +102,28 @@ export class IndexedTree implements MerkleTree {
   }
 
   /**
+   * TODO: Just export this or keep it static?
+   * Returns an empty leaf of the tree.
+   * @returns An empty leaf.
+   */
+  static initialLeaf(): LeafData {
+    return initialLeaf;
+  }
+
+  /**
    * Returns the root of the tree.
    * @returns The root of the tree.
    */
   public getRoot(includeUncommitted: boolean): Buffer {
     return this.underlying.getRoot(includeUncommitted);
+  }
+
+  /**
+   * Returns the depth of the tree.
+   * @returns The depth of the tree.
+   */
+  public getDepth(): number {
+    return this.underlying.getDepth();
   }
 
   /**
@@ -141,147 +143,6 @@ export class IndexedTree implements MerkleTree {
     for (const leaf of leaves) {
       await this.appendLeaf(leaf);
     }
-  }
-
-  /**
-   * Each base rollup needs to provide non membership / inclusion proofs for each of the nullifiers
-   * generated in the kernel circuit that it is rolling up.
-   *
-   * As leaves are batch inserted at the end, batch updates are a special case.
-   *
-   * WARNING: This function has side effects, it will insert values into the tree.
-   *
-   * Assumptions:
-   * 1. There are 8 nullifiers provided and they are all unique
-   * 2. If kc 0 has 1 nullifier, and kc 1 has 3 nullifiers the layout will assume to be the sparse
-   *   nullifier layout: [kc0N, 0, 0, 0, kc1N, kc1N, kc1N, 0]
-   *
-   * TODO: include indepth insertion writeup in this comment
-   * @param leaves Values to insert into the tree
-   * @returns
-   */
-  public async getAndPerformBaseRollupBatchInsertionProofs(
-    leaves: Buffer[],
-  ): Promise<LowNullifierWitnessData[] | undefined> {
-    // Keep track of the touched during batch insertion
-    const touchedNodes: Set<number> = new Set<number>();
-
-    // Return data
-    const lowNullifierWitnesses: LowNullifierWitnessData[] = [];
-    const startInsertionIndex: bigint = this.getNumLeaves();
-    let currInsertionIndex: bigint = startInsertionIndex;
-
-    // Leaf data of hte leaves to be inserted
-    const insertionSubtree: LeafData[] = [];
-
-    // Low nullifier membership proof sibling paths
-    for (const leaf of leaves) {
-      const newValue = toBigIntBE(leaf);
-      const indexOfPrevious = this.findIndexOfPreviousValue(newValue);
-
-      // NOTE: null values for nullfier leaves are being changed to 0n current impl is a hack
-      // Default value
-      const nullifierLeaf: LeafData = {
-        value: newValue,
-        nextIndex: 0n,
-        nextValue: 0n,
-      };
-
-      if (touchedNodes.has(indexOfPrevious.index)) {
-        // If the node has already been touched, then we return an empty leaf and sibling path
-        const emptySP = new SiblingPath();
-        emptySP.data = Array(this.underlying.getDepth()).fill(
-          Buffer.from('0000000000000000000000000000000000000000000000000000000000000000', 'hex'),
-        );
-        const witness: LowNullifierWitnessData = {
-          preimage: initialLeaf,
-          index: 0n,
-          siblingPath: emptySP,
-        };
-        lowNullifierWitnesses.push(witness);
-      } else {
-        // If the node has not been touched, we update its low nullifier pointer, but we do NOT insert it yet, inserting it now
-        // will alter non membership paths of the not yet inserted members
-        // Insertion is done at the end once updates have already occurred.
-        touchedNodes.add(indexOfPrevious.index);
-
-        const lowNullifier = this.getLatestLeafDataCopy(indexOfPrevious.index);
-
-        // If no low nullifier can be found, abort - this means the nullifier is invalid
-        // in some way (it should not happen)
-        if (lowNullifier === undefined) {
-          return undefined;
-        }
-
-        // Get sibling path for existence of the old leaf
-        const siblingPath = await this.underlying.getSiblingPath(BigInt(indexOfPrevious.index));
-
-        // Update the running paths
-        const witness = {
-          preimage: lowNullifier,
-          index: BigInt(indexOfPrevious.index),
-          siblingPath: siblingPath,
-        };
-        lowNullifierWitnesses.push(witness);
-
-        // Update subtree insertion leaf from null data
-        nullifierLeaf.nextIndex = lowNullifier.nextValue;
-        nullifierLeaf.nextValue = lowNullifier.nextValue;
-
-        // Update the current low nullifier
-        lowNullifier.nextIndex = currInsertionIndex;
-        lowNullifier.nextValue = BigInt(newValue);
-
-        // Update the old leaf in the tree
-        this.cachedLeaves[Number(indexOfPrevious.index)] = lowNullifier;
-        await this.underlying.updateLeaf(
-          hashEncodedTreeValue(lowNullifier, this.hasher),
-          BigInt(indexOfPrevious.index),
-        );
-      }
-
-      // increment insertion index
-      currInsertionIndex++;
-      insertionSubtree.push(nullifierLeaf);
-    }
-
-    // Create insertion subtree and forcefully insert in series
-    // Here we calculate the pointers for the inserted values, if they have not already been updated
-    for (let i = 0; i < leaves.length; i++) {
-      const newValue = toBigIntBE(leaves[i]);
-
-      // We have already fetched the new low nullifier for this leaf, so we can set its low nullifier
-      const lowNullifier = lowNullifierWitnesses[i].preimage;
-      // If the lowNullifier is 0, then we check the previous leaves for the low nullifier leaf
-      if (lowNullifier.value === 0n && lowNullifier.nextIndex === 0n && lowNullifier.nextValue === 0n) {
-        for (let j = 0; j < i; j++) {
-          if (
-            (insertionSubtree[j].nextValue > newValue && insertionSubtree[j].value < newValue) ||
-            (insertionSubtree[j].nextValue == 0n && insertionSubtree[j].nextIndex == 0n)
-          ) {
-            insertionSubtree[j].nextIndex = startInsertionIndex + BigInt(i);
-            insertionSubtree[j].nextValue = newValue;
-          }
-        }
-      }
-    }
-
-    // For each calculated new leaf, we insert it into the tree at the next position
-    for (let i = 0; i < leaves.length; i++) {
-      // We can skip inserting empty leaves
-      if (
-        !(
-          insertionSubtree[i].value === 0n &&
-          insertionSubtree[i].nextIndex === 0n &&
-          insertionSubtree[i].nextValue === 0n
-        )
-      ) {
-        this.cachedLeaves[Number(startInsertionIndex) + i] = insertionSubtree[i];
-        await this.underlying.appendLeaves([hashEncodedTreeValue(insertionSubtree[i], this.hasher)]);
-      }
-    }
-
-    return lowNullifierWitnesses;
   }
 
   /**
@@ -310,6 +171,17 @@ export class IndexedTree implements MerkleTree {
    */
   public async getSiblingPath(index: bigint, includeUncommitted: boolean): Promise<SiblingPath> {
     return await this.underlying.getSiblingPath(index, includeUncommitted);
+  }
+
+  /**
+   * Exposes the underlying tree's update leaf method
+   * @param leaf - The hash to set at the leaf
+   * @param index - The index of the element
+   */
+  public async updateLeaf(leaf: LeafData, index: bigint): Promise<void> {
+    this.cachedLeaves[Number(index)] = leaf;
+    const encodedLeaf = encodeTreeValue(leaf);
+    await this.underlying.updateLeaf(encodedLeaf, index);
   }
 
   /**
