+++ yarn-project/world-state/src/world-state-db/merkle_trees.ts
@@ -192,6 +192,16 @@ export class MerkleTrees implements MerkleTreeDb {
   }
 
   /**
+   * Updates a leaf in a tree at a given index.
+   * @param treeId - The ID of the tree
+   * @param leaf - The new leaf value
+   * @param index - The index to insert into
+   */
+  public async updateLeaf(treeId: IndexedMerkleTreeId, leaf: Buffer | LeafData, index: bigint): Promise<void> {
+    return await this.synchronise(() => this.trees[treeId].updateLeaf(leaf, index));
+  }
+
+  /**
    * Waits for all jobs to finish before executing the given function.
    * @param fn - The function to execute.
    * @returns Promise containing the result of the function.
@@ -210,6 +220,7 @@ export class MerkleTrees implements MerkleTreeDb {
       treeId,
       root: this.trees[treeId].getRoot(includeUncommitted),
       size: this.trees[treeId].getNumLeaves(includeUncommitted),
+      depth: this.trees[treeId].getDepth(),
     } as TreeInfo;
     return Promise.resolve(treeInfo);
   }
